REF: refactor GenerateConstructor and improve its UI titles
