+++ src/main/resources/META-INF/rust-core.xml
@@ -1080,9 +1080,9 @@
             <reference id="Rust.ShowSingleStepMacroExpansionAction"/>
         </group>
 
-        <action class="org.rust.ide.refactoring.generateConstructor.GenerateConstructorAction"
+        <action class="org.rust.ide.refactoring.generate.constructor.GenerateConstructorAction"
                 id="Rust.GenerateConstructor"
-                text="Generate Constructor">
+                text="Constructor">
             <add-to-group group-id="GenerateGroup"/>
         </action>
